refactor and add formatDate
