+++ src/helpers/formatters.js
@@ -3,8 +3,17 @@ import MomentUtils from '@date-io/moment'
 const moment = new MomentUtils()
 
 const DEFAULT_DATETIME_FORMAT = 'YYYY-MM-DD HH:mm'
+const DEFAULT_DATE_FORMAT = 'YYYY-MM-DD'
 
 export const formatDateTime = (timestamp, format = DEFAULT_DATETIME_FORMAT) => {
+	if (!timestamp) {
+		return 'no date'
+	}
+	const date = moment.date(timestamp)
+	return date.format(format)
+}
+
+export const formatDate = (timestamp, format = DEFAULT_DATE_FORMAT) => {
 	if (!timestamp) {
 		return 'no date'
 	}
